pyproject.toml
[project]
name = "web-hw-12"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi>=0.135.1",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
